// web/src/lib/api/generated/models/githubComNgodingvarengMemoriaInternalDeliveryRestDtoCircleResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Book API
 * OpenAPI spec version: 1.0
 */

export interface GithubComNgodingvarengMemoriaInternalDeliveryRestDtoCircleResponse {
  color_hex?: string;
  created_at?: string;
  created_by_user_id?: string;
  description?: string;
  dissolved_at?: string;
  id?: string;
  image_path?: string;
  name?: string;
  updated_at?: string;
}